+++ changelog.md
@@ -9,6 +9,7 @@
 		* Disabled features will not appear.
 		* Mod compat that isn't enabled (either due to absence of mod or otherwise) will not appear.
 		* Y-values for ores matches what is configured
+* ADDED: NuclearCraft compatibility
 * CHANGED: Updated Field Manual to correspond with the actual Pro Pick's featureset
 * CHANGED: How ore generation is done - should massively reduce or eliminate the amount of cascading world generation (Thanks to thiakil <3)
 * CHANGED: How Geolosys saves some world data - this is now done on another thread (another thanks to thiakil <3)